configure: Try to find a versioned Python2 interpreter.
Newer distributions, esp. Fedora 31+, changed the unversioned
symlink to the Python interpreter from Python2 to Python3.

We are now facing problems, because the AM_PATH_PYTHON macro
tries the unversioned python symlink first to find the Python2
interpreter, thus we need to find and specify the full path to
the system's Python2 interpreter before calling this macro.