Fix 100% CPU usage (again) due to log rotate
If the num_logs is set to 0 and keep_logs is set, we go back into a
loop of MAX_INT! in rotate_logs().

commit 9145e97c ("Do not rotate logs when num_logs < 2.") fixed the
issue with not going further if num_logs is less than 2, because if
num_logs is zero, we trigger this bug because of the loop:

  for (i=num_logs - 1; i>1; i--) {

As i is an unsigned int, if num_logs is zero, we initialize i to
"0 - 1" or 4,294,967,295. Thus, runs this loop over 4 billion times!

But this caused a regressing if keep_logs is set, so this
num_logs < 2 was skipped if keep_logs is set, causing the huge loop
to run once again if num_logs is zero!

There's no reason i needs to be unsigned, if i is signed, then if we
pass in num_logs = 0, then i will start off as -1, and -1 > 1 will fail
the loop and fix the issue. I don't envision anyone wanting to
keep 2,147,483,648 log files around.

Fixes: a7f9f8b5 ("Fix auditd regression where keep_logs is limited by rotate_logs 2 file test")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>