auditctl: support configuring audit_backlog_warn_time
The kernel audit subsystem has support for configuring
a variable named audit_backlog_warn_time.  This commit
adds support to auditctl for getting and setting the
value of audit_backlog_warn_time.

If audit_backlog_warn_time is greater than zero and if,
after sleeping while the length of the audit queue exceeds
the audit_backlog_limit, the total time spent sleeping is
greater than or equal to audit_backlog_warn_time, then a
warning is printed with the total time spent sleeping and
the value of audit_backlog_warn_time.

Signed-off-by: Max Englander <max.englander@gmail.com>