Do not malloc name/value pairs in nvlist
This patch changes nvlist to have one copy of the record which is
parsed up into pointers. These pointers all point to different
addresses in the record. This way we do not have to malloc each and
every name/value pair and then subsequently free them.

By doing this, we reduce malloc from 55 million times down to 5 million
times in a simple test with ausearch --format csv. While that may vary
on other systems, the ratio should be about the same.