Hookup auditctl with io_uring
This patch adds the ability to insert an io_uring rule into the
kernel. Because there is a no kernel feature to determine at run
time if the kernel supports io_uring auditing, we add a configure
switch --with-io_uring so distributions can decide if the kernel
supports it or not.