Configure fix (#304)
* Remove unnecessary double quotation mark from configure.ac

Some text editors do not expect a quotation mark in AC_MSG_ERROR
and in similar macros. Due to that, syntax highlighting is ruined
over the rest of the file.

* Add a check for the swig program if either --with-python or
--with-python3 is specified

In case one of the above written configuration options was
specifed and the swig program was not present, compilation failed
due to not having swig installed.