Change the audisp af_unix builtin plugin to external program
The af_unix plugin allows audit events to be sent to an af_unix socket
for other programs to use. The facility was moved from audispd into
auditd during the 3.0 development work. The bahavior of audisp was turned
into a thread.

The af_unix plugin has the behavior of adding a libev watcher for incoming
connections and then deleting the watcher on disconnect. Libev is not
multithread safe and eventually a global variables, fdchanges and fdchangecnt,
get into a state where no more connection can be made.

To fix this, the builtin code was moved into an external program so that
libev is not involved. See issue #300 for more informantion.