Adjust rules that have watches
This patch adjusts all the rules by factoring out any watch in the
old '-w' notation. In cases where watches are in the newer syscall
format, it adds an arch field. Because all watches are now arch dependent,
we have to double all watches so they are b32 and b64.