audispd: Variable amount of arguments
Every plugin configuration line comprises a key = value pair,
optionally followed by an additional option at the line's end.
Previously, if the value consisted of multiple words, an error
would occur. This commit addresses the issue, allowing for
the specification of multiple space-separated arguments in a
plugin configuration. For instance, the audisp-syslog plugin
can handle 3 arguments, but we were not able to pass all 3
at a single time due to this limitation.