ausearch format: Fix display of renamed file (#411)
In some cases, ausearch was not correctly showing
the new name of a renamed file when searching for
audit events. If the target file didn’t exist prior
to the rename, ausearch was unable to parse the new
file name. This occurred because ausearch attempted
to retrieve this information from the 7th record,
which is absent when the target file does not exist.