Do not call check_space_left() at startup.
During boot and auditd initialization, auditd is not
fully initialized and operational. This means that when
calling init_event(), we can't notify the parent process
that initialization is complete. However, if space_left_action
is set to SINGLE, we must avoid switching to that runlevel (during boot).
Before entering the SINGLE runlevel, it requires auditd to finish
initialization. But auditd will not start properly or signal
the init system that it has started, as it is blocked by
the attempt to switch to single-user mode, resulting in a
deadlock.