Log runlevel changes to console (during boot) (#413)
* Add capability to log runlevel changes to console

Audit may initiate runlevel changes in several scenarios,
such as when available disk space for audit logs becomes
critically low. In these situations, audit can halt the
system during boot, but without clear evidence indicating
that audit was responsible for the shutdown. With the patch
applied, audit will write to /dev/console when doing runlevel
changes.

* Add printf function attribute to write_to_console prototype

If runlevel is not HALT, we will wait until initd returns an
exit code. After that we can display to console if it succeeded.

* Do not call check_space_left() at startup.

During boot and auditd initialization, auditd is not
fully initialized and operational. This means that when
calling init_event(), we can't notify the parent process
that initialization is complete. However, if space_left_action
is set to SINGLE, we must avoid switching to that runlevel (during boot).
Before entering the SINGLE runlevel, it requires auditd to finish
initialization. But auditd will not start properly or signal
the init system that it has started, as it is blocked by
the attempt to switch to single-user mode, resulting in a
deadlock.