Move the time parser to central location
The time string parser was originally used in the ids component.
Moving it to a central location so that other components can use it.
It takes a string as input and returns the number of seconds.

The time string can take the form of: 10m, 1h, 2d, 6m to mean
10 minutes, 1 hour, 2 days, and 6 months respectively. Anything
entirely numeric is in seconds.