Optimize event formatting
In auditd-events.c, format_raw knows the length of the format_buffer.
Fix the function to remember and return the length so we can get rid of
a call to strlen. Optimize trimming newlines in format_raw. Turns out
auparse makes its own copy of a buffer when using AUSOURCE_BUFFER. So,
stop making a copy with strdup. This reduces memory churn.