Add enriched record test
This test is to ensure that when auditd is configured to create
enriched events, that it does so correctly. Correct is defined as:

enriched length > raw length
AUDIT_INTERP_SEPARATOR is found in the record
A capitalized AUID field is found in the record

The actual test is not being enabled just yet.