Make af_unix plugin ignore SIGTERM if not from auditd
This reverts the previous patch. It wouldn't allow for the transfer of
DAEMON_END to any pluugin. That can be a problem for remote logging.
Instead what this will do is look to see who sent the signal. If its
not from ppid, then ignore it under the asumption that the dispatcher
will be sending the signal soon.