aureport/ausearch: check for non-input stdin pipes
Calling aureport or ausearch on a remote host e.g.:

  # ssh host aureport

will make aureport hang on read() because stdin is seen as a pipe (from
SSH).  This can be worked around with "ssh -t", but the aforementioned
behaviour is not expected anyway.

Fix this by checking if stdin is a pipe, and, if so, poll it to check
for available data to read, which will return false for the reproducer.

Following examples now works, or continue to work, successfully:

  # aureport
  # cat audit.log | aureport
  # ssh host aureport
  # ssh host "cat audit.log | aureport"
  # cat audit.log | ssh host aureport

Signed-off-by: Enzo Matsumiya <ematsumiya@suse.de>