Add support for file_getattr and file_setattr syscalls
The file_getattr()/file_setattr() are used to set extended file
attributes. These syscalls take dirfd in conjunction with the pathname
argument. The syscall then operates on inode opened according to
openat(2) semantics.  This is an alternative to
FS_IOC_FSGETXATTR/FS_IOC_FSSETXATTR ioctl with a difference that file
don't need to be open as file can be referenced with a path instead of
fd.

Source: https://lwn.net/Articles/1027795/