Deprecate the idea of plugins being anything but outbound
At one point a long time ago, there was an idea that plugins
could be chained together to create a pipeline. it was never
implemented and no one is asking for that. So, just get rid
of the whole notion. It will still parse the direction
config directive but will unconditionally set it to out.
Nothing in audisp uses this information.