Disable some tests when static is disabled
When static is disabled, it prevents a few tests from working in the
auparse tests. This is because of visibility not working.

Added a new BUILD_STATIC conditional in the build scripts to detect
when static libraries are unavailable during configuration.

Updated the auparse test makefile to exclude internal tests by default
and only include databuf_test, lru_cache_test, and uid_name_wrap_test
when static linking is enabled.