Fix ASAN usage
Previously if ASAN libraries were detected, it would enable the use
of ASAN in the auparse unit tests. This means it can be enabled by accident
when people didn't want it at all.

Now require --with-asan to enable ASAN for unit tests only. And support
ASAN in all unit tests.